En-2311 Updates package from install to job and database table
