--- plugins/entando-plugin-jpdigital-exchange/src/main/java/org/entando/entando/aps/system/services/digitalexchange/install/JobType.java
@@ -1,6 +1,0 @@
-package org.entando.entando.aps.system.services.digitalexchange.install;
-
-public enum JobType {
-    INSTALL,
-    UNINSTALL
-}
