+++ plugins/entando-plugin-jpdigital-exchange/src/main/java/org/entando/entando/aps/system/services/digitalexchange/job/JobExecutionException.java
@@ -11,17 +11,15 @@
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
  */
-package org.entando.entando.aps.system.services.digitalexchange.install;
+package org.entando.entando.aps.system.services.digitalexchange.job;
 
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.DataBinder;
+public class JobExecutionException extends RuntimeException {
 
-/**
- * Utility methods used to register custom SpEL functions.
- */
-public class CommandUtils {
+    public JobExecutionException(String message) {
+        super(message);
+    }
 
-    public static BindingResult getBindingResult() {
-        return new DataBinder(null).getBindingResult();
+    public JobExecutionException(String message, Throwable cause) {
+        super(message, cause);
     }
 }
