+++ plugins/entando-plugin-jpdigital-exchange/src/main/java/org/entando/entando/aps/system/services/digitalexchange/job/JobType.java
@@ -1,0 +1,6 @@
+package org.entando.entando.aps.system.services.digitalexchange.job;
+
+public enum JobType {
+    INSTALL,
+    UNINSTALL
+}
