+++ plugins/entando-plugin-jpdigital-exchange/src/main/java/org/entando/entando/web/digitalexchange/install/DigitalExchangeInstallResource.java
@@ -21,7 +21,7 @@ import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import java.net.URISyntaxException;
 import javax.servlet.http.HttpServletRequest;
-import org.entando.entando.aps.system.services.digitalexchange.install.DigitalExchangeJob;
+import org.entando.entando.aps.system.services.digitalexchange.job.DigitalExchangeJob;
 import org.entando.entando.web.common.annotation.RestAccessControl;
 import org.entando.entando.web.common.model.SimpleRestResponse;
 import org.springframework.http.MediaType;
