+++ plugins/entando-plugin-jpdigital-exchange/src/main/java/org/entando/entando/web/digitalexchange/install/DigitalExchangeInstallResourceController.java
@@ -17,9 +17,9 @@ import com.agiletec.aps.system.services.user.UserDetails;
 import java.net.URI;
 import java.net.URISyntaxException;
 import javax.servlet.http.HttpServletRequest;
-import org.entando.entando.aps.system.services.digitalexchange.install.DigitalExchangeJob;
-import org.entando.entando.aps.system.services.digitalexchange.install.DigitalExchangeComponentInstallationService;
-import org.entando.entando.aps.system.services.digitalexchange.install.JobType;
+import org.entando.entando.aps.system.services.digitalexchange.job.DigitalExchangeJob;
+import org.entando.entando.aps.system.services.digitalexchange.job.DigitalExchangeComponentInstallationService;
+import org.entando.entando.aps.system.services.digitalexchange.job.JobType;
 import org.entando.entando.web.common.model.SimpleRestResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
