+++ plugins/entando-plugin-jpdigital-exchange/src/test/java/org/entando/entando/aps/system/services/digitalexchange/job/ComponentZipUtil.java
@@ -11,7 +11,7 @@
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
  */
-package org.entando.entando.aps.system.services.digitalexchange.install;
+package org.entando.entando.aps.system.services.digitalexchange.job;
 
 import java.io.File;
 import java.io.FileOutputStream;
