+++ plugins/entando-plugin-jpdigital-exchange/src/test/java/org/entando/entando/web/digitalexchange/install/DigitalExchangeInstallResourceIntegrationTest.java
@@ -28,8 +28,8 @@ import org.entando.entando.aps.system.init.IInitializerManager;
 import org.entando.entando.aps.system.init.model.SystemInstallationReport;
 import org.entando.entando.aps.system.services.digitalexchange.client.DigitalExchangeOAuth2RestTemplateFactory;
 import org.entando.entando.aps.system.services.digitalexchange.client.DigitalExchangesMocker;
-import org.entando.entando.aps.system.services.digitalexchange.install.ComponentZipUtil;
-import org.entando.entando.aps.system.services.digitalexchange.install.JobType;
+import org.entando.entando.aps.system.services.digitalexchange.job.ComponentZipUtil;
+import org.entando.entando.aps.system.services.digitalexchange.job.JobType;
 import org.entando.entando.aps.system.services.group.IGroupService;
 import org.entando.entando.aps.system.services.label.ILabelService;
 import org.entando.entando.aps.system.services.pagemodel.IPageModelService;
@@ -55,8 +55,8 @@ import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.test.web.servlet.ResultActions;
-import org.entando.entando.aps.system.services.digitalexchange.install.DigitalExchangeJob;
-import org.entando.entando.aps.system.services.digitalexchange.install.JobStatus;
+import org.entando.entando.aps.system.services.digitalexchange.job.DigitalExchangeJob;
+import org.entando.entando.aps.system.services.digitalexchange.job.JobStatus;
 import org.entando.entando.aps.system.services.storage.IStorageManager;
 import org.entando.entando.web.common.model.SimpleRestResponse;
 import org.junit.After;
